# swagger.yaml
openapi: 3.0.0
info:
  title: Intelligent Vendor Routing Platform
  description: API documentation for the intelligent vendor routing and failover system.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /route:
    post:
      summary: Route request
      description: Main endpoint. Routes the client request to the best vendor based on configured routing strategies and failover logic.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - capability
              properties:
                capability:
                  type: string
                  example: PAN_VERIFICATION
                payload:
                  type: object
                  example: {"pan_number": "ABCDE1234F"}
                requiredFeatures:
                  type: array
                  items:
                    type: string
                  example: ["name_match"]
      responses:
        '200':
          description: Request successfully routed and processed by a vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '503':
          description: All vendors failed or are unavailable
...
  /routing-logs:
    get:
      summary: Get routing logs
      description: Retrieves the history of routing decisions, failovers, and skip reasons.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Number of logs to return
        - in: query
          name: capability
          schema:
            type: string
          description: Filter logs by capability
        - in: query
          name: vendor
          schema:
            type: string
          description: Filter logs by vendor used
        - in: query
          name: status
          schema:
            type: string
            enum: [SUCCESS, FAILED]
          description: Filter logs by routing status
      responses:
        '200':
          description: Routing logs
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
...
    RouteResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
          nullable: true
        metadata:
          type: object
          properties:
            request_id:
              type: string
            capability:
              type: string
            vendor_used:
              type: string
            strategy:
              type: string
            latency_ms:
              type: integer
            skipped_vendors:
              type: array
              items:
                type: object
            failover_chain:
              type: array
              items:
                type: object
            timestamp:
              type: string